Refactor face utilities into module
